feat(dataDownloadListActions): Modularized code for DownloadAllFiles and DownloadAllModal
