+++ src/Discovery/DiscoveryDetails/DataDownloadList/ActionButtons/DownloadAllModal/DownloadAllModal.tsx
@@ -1,0 +1,31 @@
+import React from 'react';
+import { Button, Modal } from 'antd';
+
+const DownloadAllModal = ({ downloadStatus, setDownloadStatus }) => {
+  return (
+    <Modal
+      closable={false}
+      open={downloadStatus.message.active}
+      title={downloadStatus.message.title}
+      footer={
+        <Button
+          onClick={() =>
+            setDownloadStatus({
+              ...downloadStatus,
+              message: {
+                title: '',
+                content: <React.Fragment />,
+                active: false,
+              },
+            })
+          }
+        >
+          Close
+        </Button>
+      }
+    >
+      {downloadStatus.message.content}
+    </Modal>
+  );
+};
+export default DownloadAllModal;
